+++ SOUP.user.js
@@ -3,7 +3,7 @@
 // @namespace   https://github.com/vyznev/
 // @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
 // @author      Ilmari Karonen
-// @version     1.39.0
+// @version     1.39.1
 // @copyright   2014-2015, Ilmari Karonen (http://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; http://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -328,12 +328,6 @@ fixes.mso286009 = {
 	url:	"http://meta.stackoverflow.com/q/286009",
 	sites:	/^(meta\.)?stackoverflow\./,
 	css:	".nav.askquestion { margin-left: 26px }"
-};
-fixes.mso283939 = {
-	title:	"Flag for diamond moderator textarea breaks out of the dialog",
-	url:	"http://meta.stackoverflow.com/q/283939",
-	// should be safe to apply on all sites; sites other than SO already have similar CSS
-	css:	".mod-attention-subform textarea { font-size: 12px; width: 600px; padding: 5px }"
 };
 fixes.mse250407 = {
 	title:	"User signature cards on old revisions look funny",